refactor: standardize naming across codebase
- Rename chore.dart to chord.dart (typo fix)
- Standardize all factory constructors to .parse() instead of
  mixed .parse()/.fromXml(): ScorePartwise, PartList, MidiInstrument,
  MidiChannel, MidiProgram, MovementNumber, MovementTitle
- Keep MusicXmlDocument.fromXml since it takes XmlDocument (not XmlElement)

Made-with: Cursor
