+++ lib/src/element/part/measure/note/chord.dart
